requirements.txt
distro==1.4.0
numpy==1.17.4
pandas==0.24.2
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
tabula-py==1.4.2
